Consolidate Utils.
