+++ packages/core/internals/library-utils/package.json
@@ -1,0 +1,7 @@
+{
+	"name": "@aws-amplify/core/internals/library-utils",
+	"types": "../../../lib-esm/libraryUtils.d.ts",
+	"main": "../../../lib/libraryUtils.js",
+	"module": "../../../lib-esm/libraryUtils.js",
+	"react-native": "../../../lib-esm/libraryUtils.js"
+}
