+++ packages/core/src/Cache/BrowserStorageCache.ts
@@ -6,7 +6,7 @@ import { defaultConfig, getCurrTime } from './Utils';
 import { StorageCache } from './StorageCache';
 import { ICache, CacheConfig, CacheItem, CacheItemOptions } from './types';
 import { asserts } from '../Util/errors/AssertError';
-import { STORAGE_CACHE_EXCEPTION } from '../constants';
+import { STORAGE_CACHE_EXCEPTION } from '../Util/Constants';
 import { getCurrSizeKey } from './Utils/CacheUtils';
 
 const logger = new Logger('Cache');
