+++ packages/core/src/RNComponents/index.ts
@@ -1,7 +1,7 @@
 // Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-import { browserOrNode } from '../JS';
+import { browserOrNode } from '../Util/JS';
 import { StorageHelper } from '../StorageHelper';
 
 export const Linking = {};
