+++ packages/core/src/Util/JS.ts
