+++ packages/core/src/Util/errors/AssertError.ts
@@ -1,6 +1,6 @@
 // Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
-import { AmplifyError } from '../../Errors';
+import { AmplifyError } from '../Errors';
 import { ErrorParams } from '../../types';
 
 export function asserts(
