+++ packages/core/src/libraryUtils.ts
@@ -1,0 +1,6 @@
+// Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+// SPDX-License-Identifier: Apache-2.0
+
+/*
+This file maps top-level exports from `@aws-amplify/core/internals/library-utils`.
+*/
