+++ packages/core/src/singleton/index.ts
@@ -1,9 +1,9 @@
 // Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
 import { AuthClass } from './Auth';
-import { Hub } from '../Hub';
+import { Hub } from '../Hub/Hub';
 import { LibraryOptions, ResourcesConfig } from './types';
-import { AmplifyError } from '../Errors';
+import { AmplifyError } from '../Util/Errors';
 import { FetchAuthSessionOptions } from './Auth/types';
 
 // TODO(v6): add default AuthTokenStore for each platform
